+++ ui/listener-api/src/main/scala/pl/touk/nussknacker/ui/listener/ProcessChangeListener.scala
@@ -3,12 +3,12 @@ package pl.touk.nussknacker.ui.listener
 import scala.concurrent.ExecutionContext
 
 trait ProcessChangeListener {
-  def handle(event: ProcessChangeEvent)(implicit ec: ExecutionContext, user: ListenerUser): Unit
+  def handle(event: ProcessChangeEvent)(implicit ec: ExecutionContext, user: User): Unit
 }
 
 object ProcessChangeListener {
   def noop: ProcessChangeListener = new ProcessChangeListener {
-    override def handle(event: ProcessChangeEvent)(implicit ec: ExecutionContext, user: ListenerUser): Unit = event match {
+    override def handle(event: ProcessChangeEvent)(implicit ec: ExecutionContext, user: User): Unit = event match {
       case _ => ()
     }
   }