+++ ui/server/src/test/scala/pl/touk/nussknacker/ui/api/helpers/TestProcessChangeListener.scala
@@ -1,11 +1,11 @@
 package pl.touk.nussknacker.ui.api.helpers
 
-import pl.touk.nussknacker.ui.listener.{ProcessChangeEvent, ProcessChangeListener, ListenerUser}
+import pl.touk.nussknacker.ui.listener.{ProcessChangeEvent, ProcessChangeListener, User}
 
 import scala.concurrent.ExecutionContext
 
 class TestProcessChangeListener extends ProcessChangeListener {
-  override def handle(event: ProcessChangeEvent)(implicit ec: ExecutionContext, user: ListenerUser): Unit = TestProcessChangeListener.add(event)
+  override def handle(event: ProcessChangeEvent)(implicit ec: ExecutionContext, user: User): Unit = TestProcessChangeListener.add(event)
 }
 
 object TestProcessChangeListener {